#31 Reviewed the code - refactoring - only unimportant changes.
The open problem with this example is that there is hard-coded reference to resources, i.e. an application that is ready for localization is separated into two conceptual blocks, a block that contains all user interface elements and a block that contains executable code. In fact it is but tightly coupled, i.e. the resources must be compiled before the library.

I have added TPA Localizability example  #34 to describe this drawback.

Anyway for the sake of simplicity this example works as is and there is no need for improvements.